feat(execution): implement ErrorHandler for improved error processing and diagnostics in Lua VM
